+++ src/jrd/err.cpp
@@ -345,7 +345,7 @@ bool ERR_post_warning(ISC_STATUS status, ...)
 
 			case isc_arg_string:
 				q = reinterpret_cast<ISC_STATUS*>(va_arg(args, TEXT*));
-				if (strlen((TEXT *) q) >= MAX_ERRSTR_LEN) {
+				if (strlen((TEXT *) q) >= (size_t)MAX_ERRSTR_LEN) {
 					status_vector[(indx - 1)] = isc_arg_cstring;
 					status_vector[indx++] = MAX_ERRSTR_LEN;
 				}
